# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/encargados": {
            "get": {
                "tags": [
                    "Encargado"
                ],
                "summary": "Get a listing of the encargado",
                "operationId": "e397d61cbb18be4bd2bc368764118882",
                "parameters": [
                    {
                        "name": "apellido",
                        "in": "query",
                        "description": "Filter by encargado apellido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paginacion",
                        "in": "query",
                        "description": "Number of rows per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/encargados/store": {
            "post": {
                "tags": [
                    "Encargado"
                ],
                "summary": "Create a new encargado",
                "description": "Store a newly created client in storage.",
                "operationId": "471776a022f331bb9e4665c7b94dc86e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "cliente_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "perez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/encargados/{id}/show": {
            "get": {
                "tags": [
                    "Encargado"
                ],
                "summary": "Get a specific client by ID",
                "description": "Display the specified client.",
                "operationId": "f08f210b40225ef3da748b54bd4dd478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the encargado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "perez"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "encargado not found"
                    }
                }
            }
        },
        "/api/encargados/{encargado}/update": {
            "put": {
                "tags": [
                    "Encargado"
                ],
                "summary": "Update a client",
                "operationId": "3e1c73030f3804233441dc0c0bf1f28c",
                "parameters": [
                    {
                        "name": "encargado",
                        "in": "path",
                        "description": "ID of the encargado to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "cliente_id": {
                                        "type": "id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}